backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
sqlalchemy==2.0.19
pydantic==2.0.3
pymysql==1.1.0
alembic==1.11.1
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
fastapi-pagination==0.12.5
